MySecondCSApp : 07
In this activity i created one comboBox and textBox.
The idea are print little message ("the Choosen colour is: ") and
acontinuation show the name from the colour selected.

Unlike the last project forms i optimitzed the code with this function:

		String res = "The choosen color is: ";
        textBox1.Text = res + comboBox1.SelectedItem.ToString();

Also in this code for more optimization i create the objects string from comboBox
at the start the app in to the 'public Form1()':

		InitializeComponent();
        comboBox1.Items.Add("red");
        comboBox1.Items.Add("purple");
        comboBox1.Items.Add("blue");
        comboBox1.Items.Add("yellow");
        comboBox1.Items.Add("cyan");