MySecondCSApp : 09
Created project form completed.

first one i created two array Strings with values (even, odd).
The buttons 'even' and 'odd' load the array string to the comboBox.
For control up to the comboBox only one time i created a condition with 'if'.
For example when we load the even array strings to comboBox first one check if vale "Nº 1" exists into the comboBox, if don't exist the array String ups to the comboBox.

Same proces with second button with numbers odd.

finally we convert the selected item into array to string, and print the number on the textBox.